Made changes to login form. Also changed file structure to create different base layouts (no navbar and footer on login page)
