+++ src/fe/app/(main)/doctrines/page.tsx
